add chef module
